Properly fixed the buffer related panic
- Pattern 1: the idt_init() called kzalloc() before memory is initialized,
  kzalloc() returned 0x0. Used physical address 0x0 as _keybuf. (Causing undefined behavior)
- Pattern 2: used a local variable (in the stack) of idt_init() as _keybuf.
  This could work in "single thread", however, when using an eventloop, idt_init()
  will return while the eventloop keeps running. This is "Use-After-Free", because the stack
  exist no more as soon as the idt_init() returns. Subsequent write to the address corrupt stack
  thus cause panic.
- Fixed by assigning the variable in file scope -> _keybuf goes to the .data section

+++ src/idt/idt.h
@@ -43,7 +43,7 @@ void idt_int_default_handler(uint32_t interrupt_number, uintptr_t frame);
 void __int21h_buffed();
 void int21h_handler(uint8_t scancode);
 void int2ch(void);
-void int2ch_handler(uint16_t scancode);
+void int2ch_handler(uint8_t scancode);
 extern FIFO8 keybuf;
 extern FIFO8 mousebuf;
